+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [v2.0.12](https://github.com/in2workspace/in2-verifier-api/releases/tag/v2.0.12)
+### Changed
+- Read bitstring-encoded lists using MSB-first ordering.
+
+## [v2.0.11](https://github.com/in2workspace/in2-verifier-api/releases/tag/v2.0.11)
+### Added
+- Add support for BitstringStatusListEntry credential status type.
+
 ## [v2.0.10](https://github.com/in2workspace/in2-verifier-api/releases/tag/v2.0.10)
 ### Added
 - Added support for cryptographic binding